OZ_OPENGL3, memset 0xee on free, water depth fixed
- clang compatibility except for defaulted functions
- use of unsigned types only where really neccessary
- many small fixes
- oz
  * unless NDEBUG defined, all freed memoriy is overwritten by 0xee's
- matrix
  * better water depth calculation
  * disabling of floating objects
- client
  * can switch between OpenGL 2.1 + EXT and OpenGL 3.0 with OZ_OPENGL3
  * OZ_BUMPMAP to enable bumpmap-able vertex format
  * better way to determine number of mipmaps, mipmaps are only generated when
    a filter that needs them is used
  * OZ_TEXTURE_COMPRESSION to enable/disable S3TC
- glsl
  * switched to GLSL 1.20 for compatibility

+++ src/oz/Buffer.cpp
@@ -101,7 +101,7 @@ namespace oz
       alloc( int( fileStat.st_size ) );
     }
 
-    int blocksRead = int( fread( data, fileStat.st_size, 1, handle ) );
+    int blocksRead = int( fread( data, size_t( fileStat.st_size ), 1, handle ) );
     fclose( handle );
 
     if( blocksRead != 1 && fileStat.st_size != 0 ) {
@@ -119,7 +119,7 @@ namespace oz
       return false;
     }
 
-    int result = int( fwrite( data, size, 1, handle ) );
+    int result = int( fwrite( data, size_t( size ), 1, handle ) );
     fclose( handle );
 
     return result == 1;
